use different fullname for v1 v2 command in test
